[#256] Refactor: ChatRoom Command / Query 로직 분리
