+++ src/main/java/com/clover/youngchat/domain/chatroom/service/command/ChatRoomCommandService.java
@@ -1,25 +1,16 @@
-package com.clover.youngchat.domain.chatroom.service;
+package com.clover.youngchat.domain.chatroom.service.command;
 
-
-import static com.clover.youngchat.domain.chatroom.constant.ChatRoomConstant.CHAT_ROOM_DETAIL_LIMIT_SIZE;
-import static com.clover.youngchat.domain.chatroom.constant.ChatRoomConstant.CHAT_ROOM_LIMIT_SIZE;
 import static com.clover.youngchat.domain.chatroom.constant.ChatRoomConstant.COUNT_ONE_FRIEND;
 import static com.clover.youngchat.domain.chatroom.constant.ChatRoomConstant.GROUP_CHATROOM_TITLE;
 import static com.clover.youngchat.domain.chatroom.constant.ChatRoomConstant.PERSONAL_CHATROOM_TITLE;
 import static com.clover.youngchat.global.exception.ResultCode.ACCESS_DENY;
-import static com.clover.youngchat.global.exception.ResultCode.NOT_FOUND_CHAT;
 import static com.clover.youngchat.global.exception.ResultCode.NOT_FOUND_CHATROOM;
 
-import com.clover.youngchat.domain.chat.dto.response.ChatRes;
-import com.clover.youngchat.domain.chat.repository.ChatRepository;
 import com.clover.youngchat.domain.chatroom.dto.request.ChatRoomEditReq;
 import com.clover.youngchat.domain.chatroom.dto.request.GroupChatRoomCreateReq;
 import com.clover.youngchat.domain.chatroom.dto.request.PersonalChatRoomCreateReq;
-import com.clover.youngchat.domain.chatroom.dto.response.ChatRoomAndLastChatGetRes;
-import com.clover.youngchat.domain.chatroom.dto.response.ChatRoomDetailGetRes;
 import com.clover.youngchat.domain.chatroom.dto.response.ChatRoomEditRes;
 import com.clover.youngchat.domain.chatroom.dto.response.ChatRoomLeaveRes;
-import com.clover.youngchat.domain.chatroom.dto.response.ChatRoomPaginationDetailGetRes;
 import com.clover.youngchat.domain.chatroom.dto.response.GroupChatRoomCreateRes;
 import com.clover.youngchat.domain.chatroom.dto.response.PersonalChatRoomCreateRes;
 import com.clover.youngchat.domain.chatroom.entity.ChatRoom;
@@ -30,7 +21,6 @@ import com.clover.youngchat.domain.user.entity.User;
 import com.clover.youngchat.domain.user.repository.UserRepository;
 import com.clover.youngchat.global.exception.GlobalException;
 import com.clover.youngchat.global.exception.ResultCode;
-import com.clover.youngchat.global.response.RestSlice;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -40,14 +30,13 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class ChatRoomService {
+@Transactional
+public class ChatRoomCommandService {
 
     private final ChatRoomRepository chatRoomRepository;
     private final ChatRoomUserRepository chatRoomUserRepository;
-    private final ChatRepository chatRepository;
     private final UserRepository userRepository;
 
-    @Transactional
     public PersonalChatRoomCreateRes createPersonalChatRoom(
         PersonalChatRoomCreateReq req, User user) {
         User friend = findByUserId(req.getFriendId());
@@ -62,7 +51,6 @@ public class ChatRoomService {
         return PersonalChatRoomCreateRes.to(chatRoom.getId(), chatRoom.getTitle());
     }
 
-    @Transactional
     public GroupChatRoomCreateRes createGroupChatRoom(GroupChatRoomCreateReq req, User user) {
         List<User> participants = new ArrayList<>(req.getFriendIds().stream()
             .map(this::findByUserId).toList());
@@ -74,7 +62,6 @@ public class ChatRoomService {
         return GroupChatRoomCreateRes.to(chatRoom.getId(), chatRoom.getTitle());
     }
 
-    @Transactional
     public ChatRoomEditRes editChatRoom(Long chatRoomId, ChatRoomEditReq req, User user) {
         ChatRoom chatRoom = findById(chatRoomId);
 
@@ -85,7 +72,6 @@ public class ChatRoomService {
         return new ChatRoomEditRes();
     }
 
-    @Transactional
     public ChatRoomLeaveRes leaveChatRoom(Long chatRoomId, User user) {
         if (!chatRoomRepository.existsById(chatRoomId)) {
             throw new GlobalException(NOT_FOUND_CHATROOM);
@@ -100,61 +86,6 @@ public class ChatRoomService {
         return new ChatRoomLeaveRes();
     }
 
-    @Transactional(readOnly = true)
-    public RestSlice<ChatRoomAndLastChatGetRes> getChatRoomList(User user, Long cursorChatId) {
-        userRepository.findById(user.getId());
-        return chatRoomUserRepository.findChatRoomsAndLastChatByUserId(user.getId(),
-            cursorChatId, CHAT_ROOM_LIMIT_SIZE);
-    }
-
-    @Transactional(readOnly = true)
-    public ChatRoomDetailGetRes getDetailChatRoom(Long chatRoomId, User user) {
-        // 채팅방에 속한 사람만 조회 가능
-        isChatRoomMember(chatRoomId, user.getId());
-        ChatRoom chatRoom = findById(chatRoomId);
-
-        List<ChatRes> chatList = chatRepository.findAllByChatRoom_Id(chatRoomId)
-            .orElseThrow(() -> new GlobalException(NOT_FOUND_CHAT))
-            .stream()
-            .map(ChatRes::to)
-            .toList();
-
-        return ChatRoomDetailGetRes.builder()
-            .title(chatRoom.getTitle())
-            .chatResList(chatList)
-            .build();
-    }
-
-    @Transactional(readOnly = true)
-//    @Cacheable(cacheNames = CHAT_ROOM_CACHE, key = "#chatRoomId.toString().concat(':').concat(#lastChatId)", cacheManager = "cacheManager")
-    public ChatRoomPaginationDetailGetRes getPaginationDetailChatRoom(Long chatRoomId,
-        Long lastChatId,
-        User user) {
-        // 채팅방에 속한 사람만 조회 가능
-        isChatRoomMember(chatRoomId, user.getId());
-        ChatRoom chatRoom = findById(chatRoomId);
-
-        RestSlice<ChatRes> chatResList =
-            chatRepository.findChatsByChatRoomId(chatRoomId, lastChatId,
-                CHAT_ROOM_DETAIL_LIMIT_SIZE);
-
-        return ChatRoomPaginationDetailGetRes.builder()
-            .title(chatRoom.getTitle())
-            .chatResList(chatResList)
-            .build();
-    }
-
-    private ChatRoom findById(Long chatRoomId) {
-        return chatRoomRepository.findById(chatRoomId).orElseThrow(() ->
-            new GlobalException(NOT_FOUND_CHATROOM));
-    }
-
-    private User findByUserId(Long userId) {
-        return userRepository.findById(userId).orElseThrow(() ->
-            new GlobalException(ResultCode.NOT_FOUND_USER));
-    }
-
-    @Transactional
     protected ChatRoom saveChatRoom(String title, List<User> participants) {
         ChatRoom chatRoom = ChatRoom.builder()
             .title(title)
@@ -168,6 +99,16 @@ public class ChatRoomService {
         chatRoomUserRepository.saveAll(chatRoomUsers);
 
         return chatRoom;
+    }
+
+    private ChatRoom findById(Long chatRoomId) {
+        return chatRoomRepository.findById(chatRoomId).orElseThrow(() ->
+            new GlobalException(NOT_FOUND_CHATROOM));
+    }
+
+    private User findByUserId(Long userId) {
+        return userRepository.findById(userId).orElseThrow(() ->
+            new GlobalException(ResultCode.NOT_FOUND_USER));
     }
 
     private void isChatRoomMember(Long chatRoomId, Long userId) {
